Move injected script
